+++ packages/af-utils-react-virtual-headless/src/models/List/index.js
@@ -8,11 +8,12 @@ import {
     DEFAULT_ESTIMATED_ITEM_SIZE,
     DEFAULT_OVERSCAN_COUNT,
     DEFAULT_ESTIMATED_WIDGET_SIZE
-} from "constants";
+} from "src/constants";
 
 import FinalResizeObserver from "../ResizeObserver";
-import call from "/utils/call";
-import Batch from "/singletons/Batch";
+import call from "src/utils/call";
+import { build, update, getLiftingLimit } from "src/utils/fTree";
+import Batch from "src/singletons/Batch";
 
 const HORIZONTAL_SCROLL_KEY = "scrollLeft";
 const VERTICAL_SCROLL_KEY = "scrollTop";
@@ -50,27 +51,6 @@ const EMPTY_TYPED_ARRAY = new TypedCache(0);
 const STICKY_HEADER_INDEX = 0;
 const STICKY_FOOTER_INDEX = 1;
 
-const buildFtree = sourceArray => {
-    const fTreeLength = sourceArray.length + 1;
-    const fTree = new TypedCache(fTreeLength);
-
-    fTree.set(sourceArray, 1);
-
-    for (let i = 1, j; i < fTreeLength; i++) {
-        j = i + (i & -i);
-        if (j < fTreeLength) {
-            fTree[j] += fTree[i];
-        }
-    }
-
-    return fTree;
-};
-
-const updateItemHeight = (fTree, i, delta, limitTreeLiftingIndex) => {
-    for (; i < limitTreeLiftingIndex; i += i & -i) {
-        fTree[i] += delta;
-    }
-};
 class List {
     horizontal = false;
     _scrollToKey = VERTICAL_SCROLL_TO_KEY;
@@ -133,10 +113,11 @@ class List {
             diff = 0,
             buff = 0,
             wasAtLeastOneSizeChanged = false,
-            lim = this.from + 1;
-
-        for (; lim < this.to; lim += lim & -lim);
-        lim = Math.min(lim, this._fTree.length);
+            lim = /*@__NOINLINE__*/ getLiftingLimit(
+                this._fTree,
+                this.from + 1,
+                this.to
+            );
 
         for (const { target } of entries) {
             index = this._elToIdx.get(target);
@@ -146,13 +127,13 @@ class List {
                 Should not take them into account.
                 This is done for performance + updateItemHeight hack would not work without it
             */
-            if (index >= this.from && index < this.to) {
+            if (index < lim) {
                 diff = target[this._sizeKey] - this._itemSizes[index];
                 if (diff) {
                     wasAtLeastOneSizeChanged = true;
                     this._itemSizes[index] += diff;
                     buff += diff;
-                    updateItemHeight(this._fTree, index + 1, diff, lim);
+                    update(this._fTree, index + 1, diff, lim);
                 }
             }
         }
@@ -161,7 +142,7 @@ class List {
             Batch._start();
 
             if (buff !== 0) {
-                updateItemHeight(this._fTree, lim, buff, this._fTree.length);
+                update(this._fTree, lim, buff, this._fTree.length);
                 this._rawScrollSize += buff;
                 this.scrollSize += buff;
                 this._run(EVT_SCROLL_SIZE);
@@ -288,7 +269,10 @@ class List {
         );
     }
 
-    _updateScrollPos = e => {
+    /*
+        "scroll" is the only event here, so switch/case is not needed.
+    */
+    handleEvent(e) {
         const curScrollPos = this._scrollPos,
             newScrollPos = this._outerNode[this._scrollKey];
 
@@ -302,7 +286,7 @@ class List {
                 this._updateRangeFromStart();
             }
         }
-    };
+    }
 
     /*
         Performs as destructor when null is passed
@@ -311,15 +295,12 @@ class List {
     setOuterNode = node => {
         if (this._outerNode) {
             this._OuterNodeResizeObserver.unobserve(this._outerNode);
-            this._outerNode.removeEventListener(
-                "scroll",
-                this._updateScrollPos
-            );
+            this._outerNode.removeEventListener("scroll", this);
         }
 
         if ((this._outerNode = node)) {
             this._OuterNodeResizeObserver.observe(node);
-            node.addEventListener("scroll", this._updateScrollPos, {
+            node.addEventListener("scroll", this, {
                 passive: true
             });
         } else {
@@ -463,11 +444,12 @@ class List {
                 );
                 this._itemSizes.set(oldItemSizes);
 
-                this._fTree = /*@__NOINLINE__*/ buildFtree(
+                this._fTree = /*@__NOINLINE__*/ build(
                     this._itemSizes.fill(
                         this._estimatedItemSize || DEFAULT_ESTIMATED_ITEM_SIZE,
                         curRowHeighsLength
-                    )
+                    ),
+                    TypedCache
                 );
             }