+++ packages/af-utils-react-virtual-headless/src/models/List/index.test.js
@@ -1,6 +1,7 @@
 import List from "./index";
 import r from "lodash/random";
-
+import clamp from "lodash/clamp";
+/*
 const ROWS_QUANTITY = 100;
 const UPDATES_QUANTITY = 1;
 
@@ -22,13 +23,22 @@ const getIndexUnoptimized = offset => {
     } while (offset >= 0);
 
     return index;
-};
+};*/
 
 describe("List model works", () => {
-    const VSList = new List();
+    const VSList = new List({ estimatedWidgetSize: 12345 });
 
-    test("Default itemCount is zero", () => {
-        expect(VSList.itemCount).toEqual(0);
+    test("Default params are ok", () => {
+        expect(VSList._scrollPos).toEqual(0);
+        expect(VSList._itemCount).toEqual(0);
+        expect(VSList.from).toEqual(0);
+        expect(VSList.to).toEqual(0);
+        expect(VSList.scrollSize).toEqual(0);
+        expect(VSList.sizesHash).toEqual(0);
+    });
+
+    test("Setting itemCount > max(int32) or < 0 throws error", () => {
+        expect(() => VSList.set({ itemCount: 0x7fffffff + 1 })).toThrow();
     });
 
     test("Calcultions with zero itemCount work correctly", () => {
@@ -36,32 +46,56 @@ describe("List model works", () => {
         expect(VSList.getOffset(0)).toEqual(0);
     });
 
-    test("Setting itemCount > max(int32) or < 0 throws error", () => {
-        expect(() => VSList._setParams(0, 2, -1, false)).toThrow();
-        expect(() => VSList._setParams(0, 2, 0x7fffffff + 1, false)).toThrow();
+    test("Default range is set correctly with 0 overscanCount", () => {
+        VSList.set({ itemCount: 100, overscanCount: 0 });
+        expect(VSList.scrollSize).toEqual(100 * VSList._estimatedItemSize);
+        expect(VSList.from).toEqual(0);
+        expect(VSList.to).toEqual(
+            clamp(
+                Math.ceil(
+                    VSList._availableWidgetSize / VSList._estimatedItemSize
+                ),
+                1,
+                VSList._itemCount
+            )
+        );
     });
 
-    test("Summation works correctly", () => {
-        VSList.setItemCount(ROWS_QUANTITY, false, () => 2);
-        VSList.setOverscan(0);
+    test("Overscan change with same itemCount does not lead to range change", () => {
+        VSList.set({ itemCount: 100, overscanCount: 500 });
+        expect(VSList.from).toEqual(0);
+        expect(VSList.to).toEqual(
+            clamp(
+                Math.ceil(
+                    VSList._availableWidgetSize / VSList._estimatedItemSize
+                ),
+                1,
+                VSList._itemCount
+            )
+        );
+    });
 
-        for (let i = 0, sum; i < UPDATES_QUANTITY; i++) {
-            sum = 0;
+    test("Default range is set correctly with non-0 overscanCount", () => {
+        VSList.set({ itemCount: 1000, overscanCount: 5 });
+        expect(VSList.scrollSize).toEqual(
+            VSList._itemCount * VSList._estimatedItemSize
+        );
+        expect(VSList.from).toEqual(0);
+        expect(VSList.to).toEqual(
+            clamp(
+                5 +
+                    Math.ceil(
+                        VSList._availableWidgetSize / VSList._estimatedItemSize
+                    ),
+                1,
+                VSList._itemCount
+            )
+        );
+    });
 
-            for (let j = 0, diff; j < ROWS_QUANTITY; j++) {
-                diff = r(5 - i, 20);
-                sum += diff;
-                VSList._updateItemHeight(j + 1, diff, VSList._fTree.length);
-                ROW_HEIGHTS[j] += diff;
-            }
-
-            for (let j = 0; j < ROWS_QUANTITY; j++) {
-                expect(VSList.getOffset(j)).toBe(getOffsetUnoptimized(j));
-            }
-
-            for (let j = 0; j < sum; j++) {
-                expect(VSList.getIndex(j)).toBe(getIndexUnoptimized(j));
-            }
+    test("Adding initial elements works correctly", () => {
+        for (let i = VSList.from; i < VSList.to; i++) {
+            VSList.el(i, document.createElement("div"));
         }
     });
 });