+++ packages/af-utils-react-virtual-headless/src/models/ResizeObserver/index.js
@@ -1,9 +1,10 @@
-const FinalResizeObserver = process.env.__IS_SERVER__
-    ? class {
-          observe() {}
-          unobserve() {}
-          disconnect() {}
-      }
-    : ResizeObserver;
+const FinalResizeObserver =
+    process.env.__IS_SERVER__ || process.env.NODE_ENV === "test"
+        ? class {
+              observe() {}
+              unobserve() {}
+              disconnect() {}
+          }
+        : ResizeObserver;
 
 export default FinalResizeObserver;