+++ packages/af-utils-react-virtual-headless/src/singletons/Batch/index.js
@@ -1,4 +1,4 @@
-import call from "/utils/call";
+import call from "src/utils/call";
 
 const Queue = new Set();