+++ packages/af-utils-styled/babel.config.js
@@ -12,7 +12,7 @@ const presets = [
     ]
 ];
 
-module.exports = {
+export default {
     plugins,
     presets
 };