+++ packages/af-utils-react-virtual-list/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @af-utils/react-virtual-list
 
+## 0.0.50
+
+### Patch Changes
+
+-   upgraded deps, changed format to esm, improved build scripts, added some tests
+-   Updated dependencies
+    -   @af-utils/react-virtual-headless@0.0.47
+    -   @af-utils/styled@0.0.19
+
 ## 0.0.49
 
 ### Patch Changes